WIP fix for Mesh_3_plugin
- io_images don't work

+++ Polyhedron/demo/Polyhedron/Plugins/Mesh_3_plugin/Volume_plane_intersection.cpp
@@ -116,7 +116,7 @@ void Volume_plane_intersection::init_buffers()
 void Volume_plane_intersection::attrib_buffers(Viewer_interface* viewer) const
 {
     QMatrix4x4 mvpMatrix;
-    double mat[16];
+    float mat[16];
     viewer->camera()->getModelViewProjectionMatrix(mat);
     for(int i=0; i < 16; i++)
     {
